+++ plugins/backend/catalog-backend-module-aws/src/providers/AWSEntityProvider.ts
@@ -20,7 +20,6 @@ import {
 } from '@backstage/plugin-catalog-node';
 import * as winston from 'winston';
 import { AccountConfig } from '../types';
-import { fromTemporaryCredentials } from '@aws-sdk/credential-providers';
 import { STS } from '@aws-sdk/client-sts';
 import {
   ANNOTATION_ORIGIN_LOCATION,
@@ -28,17 +27,20 @@ import {
 } from '@backstage/catalog-model';
 import { ANNOTATION_ACCOUNT_ID } from '../annotations';
 import { CatalogApi } from '@backstage/catalog-client';
+import { DefaultAwsCredentialsManager } from '@backstage/integration-aws-node';
+import { ConfigReader } from '@backstage/config';
+import { fromTemporaryCredentials } from '@aws-sdk/credential-providers';
+import { parse as parseArn } from '@aws-sdk/util-arn-parser';
 
 export abstract class AWSEntityProvider implements EntityProvider {
-  protected readonly accountId: string;
-  protected readonly roleArn: string;
-  private readonly externalId?: string;
-  protected readonly region: string;
+  protected readonly useTemporaryCredentials: boolean;
   protected readonly providerId?: string;
   protected readonly logger: winston.Logger;
   protected connection?: EntityProviderConnection;
   private readonly ownerTag: string | undefined;
   protected readonly catalogApi?: CatalogApi;
+  private credentialsManager: DefaultAwsCredentialsManager;
+  private account: AccountConfig;
 
   public abstract getProviderName(): string;
 
@@ -49,16 +51,26 @@ export abstract class AWSEntityProvider implements EntityProvider {
       catalogApi?: CatalogApi;
       providerId?: string;
       ownerTag?: string;
+      useTemporaryCredentials?: boolean;
     },
   ) {
-    this.accountId = account.accountId;
-    this.roleArn = account.roleArn;
-    this.externalId = account.externalId;
-    this.region = account.region;
     this.logger = options.logger;
     this.providerId = options.providerId;
     this.ownerTag = options.ownerTag;
     this.catalogApi = options.catalogApi;
+    this.account = account;
+    this.useTemporaryCredentials = !!options.useTemporaryCredentials;
+    this.credentialsManager = DefaultAwsCredentialsManager.fromConfig(
+      new ConfigReader({ aws: { accounts: [account] } }),
+    );
+  }
+
+  get accountId() {
+    return this.account.accountId;
+  }
+
+  get region() {
+    return this.account.region;
   }
 
   protected getOwnerTag() {
@@ -66,9 +78,24 @@ export abstract class AWSEntityProvider implements EntityProvider {
   }
 
   protected getCredentials() {
+    const arnParse = parseArn(this.account.roleArn ?? this.account.roleName);
+
+    const region = this.region ?? arnParse.region;
     return fromTemporaryCredentials({
-      params: { RoleArn: this.roleArn, ExternalId: this.externalId },
+      params: {
+        RoleArn: this.account.roleArn,
+        ExternalId: this.account.externalId,
+      },
+      clientConfig: region ? { region: region } : undefined,
     });
+  }
+
+  protected async getCredentialsProvider() {
+    const awsCredentialProvider =
+      await this.credentialsManager.getCredentialProvider({
+        accountId: this.accountId,
+      });
+    return awsCredentialProvider.sdkCredentialProvider;
   }
 
   protected async getGroups() {
@@ -92,13 +119,22 @@ export abstract class AWSEntityProvider implements EntityProvider {
   }
 
   protected async buildDefaultAnnotations() {
-    const sts = new STS({ credentials: this.getCredentials() });
+    const credentials = this.useTemporaryCredentials
+      ? this.getCredentials()
+      : await this.getCredentialsProvider();
+    const sts = this.useTemporaryCredentials
+      ? new STS({ credentials: credentials, region: this.region })
+      : new STS(credentials);
 
     const account = await sts.getCallerIdentity({});
 
     const defaultAnnotations: { [name: string]: string } = {
-      [ANNOTATION_LOCATION]: `${this.getProviderName()}:${this.roleArn}`,
-      [ANNOTATION_ORIGIN_LOCATION]: `${this.getProviderName()}:${this.roleArn}`,
+      [ANNOTATION_LOCATION]: `${this.getProviderName()}:${
+        this.account.roleName
+      }`,
+      [ANNOTATION_ORIGIN_LOCATION]: `${this.getProviderName()}:${
+        this.account.roleName
+      }`,
     };
 
     if (account.Account) {