+++ plugins/backend/catalog-backend-module-aws/src/providers/AWSS3BucketProvider.test.ts
@@ -33,7 +33,7 @@ const logger = createLogger({
 describe('AWSS3BucketProvider', () => {
   const config = new ConfigReader({
     accountId: '123456789012',
-    roleArn: 'arn:aws:iam::123456789012:role/role1',
+    roleName: 'arn:aws:iam::123456789012:role/role1',
     region: 'eu-west-1',
   });