+++ plugins/backend/catalog-backend-module-aws/src/providers/AWSS3BucketProvider.ts
@@ -22,7 +22,11 @@ import { AWSEntityProvider } from './AWSEntityProvider';
 import { ANNOTATION_AWS_S3_BUCKET_ARN } from '../annotations';
 import { arnToName } from '../utils/arnToName';
 import { ARN } from 'link2aws';
-import { labelsFromTags, ownerFromTags } from '../utils/tags';
+import {
+  labelsFromTags,
+  ownerFromTags,
+  relationshipsFromTags,
+} from '../utils/tags';
 import { CatalogApi } from '@backstage/catalog-client';
 
 /**
@@ -36,21 +40,32 @@ export class AWSS3BucketProvider extends AWSEntityProvider {
       catalogApi?: CatalogApi;
       providerId?: string;
       ownerTag?: string;
+      useTemporaryCredentials?: boolean;
     },
   ) {
     const accountId = config.getString('accountId');
-    const roleArn = config.getString('roleArn');
+    const roleName = config.getString('roleName');
+    const roleArn = config.getOptionalString('roleArn');
     const externalId = config.getOptionalString('externalId');
     const region = config.getString('region');
 
     return new AWSS3BucketProvider(
-      { accountId, roleArn, externalId, region },
+      { accountId, roleName, roleArn, externalId, region },
       options,
     );
   }
 
   getProviderName(): string {
     return `aws-s3-bucket-${this.accountId}-${this.providerId ?? 0}`;
+  }
+
+  private async getS3() {
+    const credentials = this.useTemporaryCredentials
+      ? this.getCredentials()
+      : await this.getCredentialsProvider();
+    return this.useTemporaryCredentials
+      ? new S3({ credentials, region: this.region })
+      : new S3(credentials);
   }
 
   async run(): Promise<void> {
@@ -64,8 +79,7 @@ export class AWSS3BucketProvider extends AWSEntityProvider {
     );
     const s3Resources: ResourceEntity[] = [];
 
-    const credentials = this.getCredentials();
-    const s3 = new S3({ credentials, region: this.region });
+    const s3 = await this.getS3();
 
     const defaultAnnotations = this.buildDefaultAnnotations();
 
@@ -92,6 +106,7 @@ export class AWSS3BucketProvider extends AWSEntityProvider {
           },
           spec: {
             owner: ownerFromTags(tags, this.getOwnerTag(), groups),
+            ...relationshipsFromTags(tags),
             type: 's3-bucket',
           },
         };