+++ plugins/backend/catalog-backend-module-aws/src/types.ts
@@ -59,7 +59,11 @@ export interface AWSAccountProviderConfig {
   /**
    * Role to assume for this account ID
    */
-  roleArn: string;
+  roleName: string;
+  /**
+   * Role ARN to assume for this account ID
+   */
+  roleArn?: string;
   /**
    * Region to use for this account ID
    */
@@ -77,7 +81,8 @@ export interface AWSAccountProviderConfig {
 
 export type AccountConfig = {
   accountId: string;
-  roleArn: string;
+  roleName: string;
+  roleArn?: string;
   externalId?: string;
   region: string;
 };