+++ plugins/frontend/backstage-plugin-iframe/src/components/types.ts
@@ -22,6 +22,7 @@ export type IFrameComponentContentProps = {
 export type IFrameComponentProps = {
   src?: string;
   srcFromAnnotation?: string;
+  templatedSrc?: string;
   title?: string;
   height?: string;
   width?: string;
@@ -30,6 +31,14 @@ export type IFrameComponentProps = {
 
 export type IFrameFromAnnotationProps = {
   srcFromAnnotation: string;
+  title?: string;
+  height?: string;
+  width?: string;
+  classes?: string;
+};
+
+export type IframeFromTemplatedSrcProps = {
+  templatedSrc: string;
   title?: string;
   height?: string;
   width?: string;